+++ R/invoke.R
@@ -1,5 +1,9 @@
 #' Invoke functions.
 #'
+#' @description
+#'
+#' \Sexpr[results=rd, stage=render]{purrr:::lifecycle("soft-deprecated")}
+#'
 #' This pair of functions make it easier to combine a function and list
 #' of parameters to get a result. `invoke` is a wrapper around
 #' `do.call` that makes it easy to use in a pipe. `invoke_map`
@@ -18,6 +22,36 @@
 #'   as `.f` the name of a function rather than its value, or as
 #'   `.x` symbols of objects rather than their values.
 #' @inheritParams map
+#'
+#' @section Life cycle:
+#'
+#' * `invoke()` is soft-deprecated as of purrr 0.3.0 and replaced by
+#'   the simpler `exec()` function reexported from rlang. `exec()`
+#'   evaluates a function call built from its inputs and supports tidy
+#'   dots:
+#'
+#'   ```
+#'   # Before:
+#'   invoke(mean, list(na.rm = TRUE), x = 1:10)
+#'
+#'   # After
+#'   exec(mean, 1:10, !!!list(na.rm = TRUE))
+#'   ```
+#'
+#' * `invoke_map()` is soft-deprecated as of purrr 0.3.0 without
+#'   replacement because it is more complex to understand than the
+#'   corresponding code using `map()`, `map2()` and `exec()`:
+#'
+#'   ```
+#'   # Before:
+#'   invoke_map(fns, list(args))
+#'   invoke_map(fns, list(args1, args2))
+#'
+#'   # After:
+#'   map(fns, exec, !!!args)
+#'   map2(fns, list(args1, args2), function(fn, args) exec(fn, !!!args))
+#'   ```
+#'
 #' @export
 #' @family map variants
 #' @examples
@@ -64,6 +98,17 @@
 #' df
 #' invoke_map(df$f, df$params)
 invoke <- function(.f, .x = NULL, ..., .env = NULL) {
+  signal_soft_deprecated(paste_line(
+    "`invoke()` is soft-deprecated as of purrr 0.3.0.",
+    "Please use `purrr::exec()` (reexported from rlang) instead.",
+    "",
+    "  # Before:",
+    "  invoke(mean, list(na.rm = TRUE), x = 1:10)",
+    "",
+    "  # After:",
+    "  exec(mean, 1:10, !!!list(na.rm = TRUE))",
+    ""
+  ))
   .env <- .env %||% parent.frame()
 
   args <- c(as.list(.x), list(...))
@@ -81,6 +126,7 @@ as_invoke_function <- function(f) {
 #' @rdname invoke
 #' @export
 invoke_map <- function(.f, .x = list(NULL), ..., .env = NULL) {
+  signal_soft_deprecated_invoke_map("", caller_env())
   .env <- .env %||% parent.frame()
   .f <- as_invoke_function(.f)
   map2(.f, .x, invoke, ..., .env = .env)
@@ -88,6 +134,7 @@ invoke_map <- function(.f, .x = list(NULL), ..., .env = NULL) {
 #' @rdname invoke
 #' @export
 invoke_map_lgl <- function(.f, .x = list(NULL), ..., .env = NULL) {
+  signal_soft_deprecated_invoke_map("_lgl", caller_env())
   .env <- .env %||% parent.frame()
   .f <- as_invoke_function(.f)
   map2_lgl(.f, .x, invoke, ..., .env = .env)
@@ -95,6 +142,7 @@ invoke_map_lgl <- function(.f, .x = list(NULL), ..., .env = NULL) {
 #' @rdname invoke
 #' @export
 invoke_map_int <- function(.f, .x = list(NULL), ..., .env = NULL) {
+  signal_soft_deprecated_invoke_map("_int", caller_env())
   .env <- .env %||% parent.frame()
   .f <- as_invoke_function(.f)
   map2_int(.f, .x, invoke, ..., .env = .env)
@@ -102,6 +150,7 @@ invoke_map_int <- function(.f, .x = list(NULL), ..., .env = NULL) {
 #' @rdname invoke
 #' @export
 invoke_map_dbl <- function(.f, .x = list(NULL), ..., .env = NULL) {
+  signal_soft_deprecated_invoke_map("_dbl", caller_env())
   .env <- .env %||% parent.frame()
   .f <- as_invoke_function(.f)
   map2_dbl(.f, .x, invoke, ..., .env = .env)
@@ -109,6 +158,7 @@ invoke_map_dbl <- function(.f, .x = list(NULL), ..., .env = NULL) {
 #' @rdname invoke
 #' @export
 invoke_map_chr <- function(.f, .x = list(NULL), ..., .env = NULL) {
+  signal_soft_deprecated_invoke_map("_chr", caller_env())
   .env <- .env %||% parent.frame()
   .f <- as_invoke_function(.f)
   map2_chr(.f, .x, invoke, ..., .env = .env)
@@ -116,6 +166,7 @@ invoke_map_chr <- function(.f, .x = list(NULL), ..., .env = NULL) {
 #' @rdname invoke
 #' @export
 invoke_map_raw <- function(.f, .x = list(NULL), ..., .env = NULL) {
+  signal_soft_deprecated_invoke_map("_raw", caller_env())
   .env <- .env %||% parent.frame()
   .f <- as_invoke_function(.f)
   map2_raw(.f, .x, invoke, ..., .env = .env)
@@ -124,6 +175,7 @@ invoke_map_raw <- function(.f, .x = list(NULL), ..., .env = NULL) {
 #' @rdname invoke
 #' @export
 invoke_map_dfr <- function(.f, .x = list(NULL), ..., .env = NULL) {
+  signal_soft_deprecated_invoke_map("_dfr", caller_env())
   .env <- .env %||% parent.frame()
   .f <- as_invoke_function(.f)
   map2_dfr(.f, .x, invoke, ..., .env = .env)
@@ -131,6 +183,7 @@ invoke_map_dfr <- function(.f, .x = list(NULL), ..., .env = NULL) {
 #' @rdname invoke
 #' @export
 invoke_map_dfc <- function(.f, .x = list(NULL), ..., .env = NULL) {
+  signal_soft_deprecated_invoke_map("_dfc", caller_env())
   .env <- .env %||% parent.frame()
   .f <- as_invoke_function(.f)
   map2_dfc(.f, .x, invoke, ..., .env = .env)
@@ -139,6 +192,27 @@ invoke_map_dfc <- function(.f, .x = list(NULL), ..., .env = NULL) {
 #' @export
 #' @usage NULL
 invoke_map_df <- invoke_map_dfr
+
+signal_soft_deprecated_invoke_map <- function(suffix, env) {
+  signal_soft_deprecated(env = env, paste_line(
+    sprintf("`invoke_map%s()` is soft-deprecated as of purrr 0.3.0.", suffix),
+    "Please use `exec()` and `map()` or `map2()` instead.",
+    "",
+    "  # Before:",
+    "  invoke_map(fns, list(args))",
+    "  invoke_map(fns, list(args1, args2))",
+    "",
+    "  # After:",
+    "  map(fns, exec, !!!args)",
+    "  map2(fns, list(args1, args2), function(fn, args) exec(fn, !!!args))"
+  ))
+
+  # Disable invoke() deprecation warning
+  scoped_options(
+    lifecycle_disable_warnings = TRUE,
+    .frame = caller_env()
+  )
+}
 
 
 #' @rdname invoke